refactor logic of friendly_id into a concern
